Attempt fixing the backup and restore scripts
This has required to eschew the `wrapProgram` script as they then go on
to figure out their "installed" location via `dirname` which now points
to the nixstore location. Thus leading to the backup being bogus and
also not written. Fortunately all of the interesting programs are shell
scripts, so we can just prepend some text to them.

Another issue was that we were deleting the `lst` file that OH was
relying on to clean up the backup afterwards.